+++ src/main/java/com/example/cosmocatsintergalacticmarketplace/service/impl/CurrencyServiceImpl.java
@@ -3,7 +3,7 @@ package com.example.cosmocatsintergalacticmarketplace.service.impl;
 import com.example.cosmocatsintergalacticmarketplace.domain.CurrencyRate;
 import com.example.cosmocatsintergalacticmarketplace.dto.CurrencyRateDto;
 import com.example.cosmocatsintergalacticmarketplace.service.CurrencyService;
-import com.example.cosmocatsintergalacticmarketplace.service.mapper.CurrencyRateMapper;
+import com.example.cosmocatsintergalacticmarketplace.web.mapper.CurrencyRateMapper;
 import org.springframework.beans.factory.annotation.Value;
 import org.springframework.stereotype.Service;
 import org.springframework.web.client.RestTemplate;
